app/requirements.txt
Flask
pymongo
markdown
requests
python-dotenv
gunicorn
Flask-Caching
redis